Change quantity to use double
Fractional quantities result from allocation
Fix test cases

+++ modules/product/src/test/java/com/opengamma/strata/product/GenericSecurityPositionTest.java
@@ -23,11 +23,11 @@ public class GenericSecurityPositionTest {
   private static final PositionInfo POSITION_INFO = PositionInfo.of(StandardId.of("A", "B"));
   private static final GenericSecurity SECURITY = GenericSecurityTest.sut();
   private static final GenericSecurity SECURITY2 = GenericSecurityTest.sut2();
-  private static final int LONG_QUANTITY = 300;
-  private static final int LONG_QUANTITY2 = 350;
-  private static final int SHORT_QUANTITY = 200;
-  private static final int SHORT_QUANTITY2 = 150;
-  private static final int QUANTITY = 100;
+  private static final double LONG_QUANTITY = 300;
+  private static final double LONG_QUANTITY2 = 350;
+  private static final double SHORT_QUANTITY = 200;
+  private static final double SHORT_QUANTITY2 = 150;
+  private static final double QUANTITY = 100;
 
   //-------------------------------------------------------------------------
   public void test_ofNet_noInfo() {
@@ -35,37 +35,37 @@ public class GenericSecurityPositionTest {
     assertEquals(test.getInfo(), PositionInfo.empty());
     assertEquals(test.getSecurity(), SECURITY);
     assertEquals(test.getLongQuantity(), QUANTITY);
-    assertEquals(test.getShortQuantity(), 0);
+    assertEquals(test.getShortQuantity(), 0d);
     assertEquals(test.getQuantity(), QUANTITY);
     assertEquals(test.getSecurityId(), SECURITY.getSecurityId());
     assertEquals(test.getCurrency(), SECURITY.getCurrency());
   }
 
   public void test_ofNet_withInfo_positive() {
-    GenericSecurityPosition test = GenericSecurityPosition.ofNet(POSITION_INFO, SECURITY, 100);
+    GenericSecurityPosition test = GenericSecurityPosition.ofNet(POSITION_INFO, SECURITY, 100d);
     assertEquals(test.getInfo(), POSITION_INFO);
     assertEquals(test.getSecurity(), SECURITY);
-    assertEquals(test.getLongQuantity(), 100);
-    assertEquals(test.getShortQuantity(), 0);
-    assertEquals(test.getQuantity(), 100);
+    assertEquals(test.getLongQuantity(), 100d);
+    assertEquals(test.getShortQuantity(), 0d);
+    assertEquals(test.getQuantity(), 100d);
   }
 
   public void test_ofNet_withInfo_zero() {
-    GenericSecurityPosition test = GenericSecurityPosition.ofNet(POSITION_INFO, SECURITY, 0);
+    GenericSecurityPosition test = GenericSecurityPosition.ofNet(POSITION_INFO, SECURITY, 0d);
     assertEquals(test.getInfo(), POSITION_INFO);
     assertEquals(test.getSecurity(), SECURITY);
-    assertEquals(test.getLongQuantity(), 0);
-    assertEquals(test.getShortQuantity(), 0);
-    assertEquals(test.getQuantity(), 0);
+    assertEquals(test.getLongQuantity(), 0d);
+    assertEquals(test.getShortQuantity(), 0d);
+    assertEquals(test.getQuantity(), 0d);
   }
 
   public void test_ofNet_withInfo_negative() {
-    GenericSecurityPosition test = GenericSecurityPosition.ofNet(POSITION_INFO, SECURITY, -100);
+    GenericSecurityPosition test = GenericSecurityPosition.ofNet(POSITION_INFO, SECURITY, -100d);
     assertEquals(test.getInfo(), POSITION_INFO);
     assertEquals(test.getSecurity(), SECURITY);
-    assertEquals(test.getLongQuantity(), 0);
-    assertEquals(test.getShortQuantity(), 100);
-    assertEquals(test.getQuantity(), -100);
+    assertEquals(test.getLongQuantity(), 0d);
+    assertEquals(test.getShortQuantity(), 100d);
+    assertEquals(test.getQuantity(), -100d);
   }
 
   public void test_ofLongShort_noInfo() {
