+++ modules/product/src/test/java/com/opengamma/strata/product/GenericSecurityTradeTest.java
@@ -22,8 +22,8 @@ public class GenericSecurityTradeTest {
   private static final TradeInfo TRADE_INFO = TradeInfo.of(date(2016, 6, 30));
   private static final GenericSecurity SECURITY = GenericSecurityTest.sut();
   private static final GenericSecurity SECURITY2 = GenericSecurityTest.sut2();
-  private static final int QUANTITY = 100;
-  private static final int QUANTITY2 = 200;
+  private static final double QUANTITY = 100;
+  private static final double QUANTITY2 = 200;
   private static final double PRICE = 123.50;
   private static final double PRICE2 = 120.50;
