+++ modules/product/src/test/java/com/opengamma/strata/product/SecurityPositionTest.java
@@ -23,11 +23,11 @@ public class SecurityPositionTest {
   private static final PositionInfo POSITION_INFO = PositionInfo.of(StandardId.of("A", "B"));
   private static final SecurityId SECURITY_ID = SecurityId.of("OG-Test", "Id");
   private static final SecurityId SECURITY_ID2 = SecurityId.of("OG-Test", "Id2");
-  private static final int LONG_QUANTITY = 300;
-  private static final int LONG_QUANTITY2 = 350;
-  private static final int SHORT_QUANTITY = 200;
-  private static final int SHORT_QUANTITY2 = 150;
-  private static final int QUANTITY = 100;
+  private static final double LONG_QUANTITY = 300;
+  private static final double LONG_QUANTITY2 = 350;
+  private static final double SHORT_QUANTITY = 200;
+  private static final double SHORT_QUANTITY2 = 150;
+  private static final double QUANTITY = 100;
 
   //-------------------------------------------------------------------------
   public void test_ofNet_noInfo() {
@@ -35,35 +35,35 @@ public class SecurityPositionTest {
     assertEquals(test.getInfo(), PositionInfo.empty());
     assertEquals(test.getSecurityId(), SECURITY_ID);
     assertEquals(test.getLongQuantity(), QUANTITY);
-    assertEquals(test.getShortQuantity(), 0);
+    assertEquals(test.getShortQuantity(), 0d);
     assertEquals(test.getQuantity(), QUANTITY);
   }
 
   public void test_ofNet_withInfo_positive() {
-    SecurityPosition test = SecurityPosition.ofNet(POSITION_INFO, SECURITY_ID, 100);
+    SecurityPosition test = SecurityPosition.ofNet(POSITION_INFO, SECURITY_ID, 100d);
     assertEquals(test.getInfo(), POSITION_INFO);
     assertEquals(test.getSecurityId(), SECURITY_ID);
-    assertEquals(test.getLongQuantity(), 100);
-    assertEquals(test.getShortQuantity(), 0);
-    assertEquals(test.getQuantity(), 100);
+    assertEquals(test.getLongQuantity(), 100d);
+    assertEquals(test.getShortQuantity(), 0d);
+    assertEquals(test.getQuantity(), 100d);
   }
 
   public void test_ofNet_withInfo_zero() {
-    SecurityPosition test = SecurityPosition.ofNet(POSITION_INFO, SECURITY_ID, 0);
+    SecurityPosition test = SecurityPosition.ofNet(POSITION_INFO, SECURITY_ID, 0d);
     assertEquals(test.getInfo(), POSITION_INFO);
     assertEquals(test.getSecurityId(), SECURITY_ID);
-    assertEquals(test.getLongQuantity(), 0);
-    assertEquals(test.getShortQuantity(), 0);
-    assertEquals(test.getQuantity(), 0);
+    assertEquals(test.getLongQuantity(), 0d);
+    assertEquals(test.getShortQuantity(), 0d);
+    assertEquals(test.getQuantity(), 0d);
   }
 
   public void test_ofNet_withInfo_negative() {
-    SecurityPosition test = SecurityPosition.ofNet(POSITION_INFO, SECURITY_ID, -100);
+    SecurityPosition test = SecurityPosition.ofNet(POSITION_INFO, SECURITY_ID, -100d);
     assertEquals(test.getInfo(), POSITION_INFO);
     assertEquals(test.getSecurityId(), SECURITY_ID);
-    assertEquals(test.getLongQuantity(), 0);
-    assertEquals(test.getShortQuantity(), 100);
-    assertEquals(test.getQuantity(), -100);
+    assertEquals(test.getLongQuantity(), 0d);
+    assertEquals(test.getShortQuantity(), 100d);
+    assertEquals(test.getQuantity(), -100d);
   }
 
   public void test_ofLongShort_noInfo() {
